Move [Span|TagMap]InScope and share between api/sdk
Removes duplication between the two projects

Fixes #333
