+++ api/src/main/java/io/opentelemetry/trace/DefaultTracer.java
@@ -16,7 +16,6 @@
 
 package io.opentelemetry.trace;
 
-import io.grpc.Context;
 import io.opentelemetry.context.Scope;
 import io.opentelemetry.context.propagation.BinaryFormat;
 import io.opentelemetry.context.propagation.HttpTextFormat;
@@ -24,6 +23,7 @@ import io.opentelemetry.context.propagation.TraceContextFormat;
 import io.opentelemetry.internal.Utils;
 import io.opentelemetry.resource.Resource;
 import io.opentelemetry.trace.unsafe.ContextUtils;
+import io.opentelemetry.trace.unsafe.SpanInScope;
 import java.util.List;
 import javax.annotation.Nullable;
 
@@ -177,24 +177,5 @@ public final class DefaultTracer implements Tracer {
     }
 
     private NoopBinaryFormat() {}
-  }
-
-  private static final class SpanInScope implements Scope {
-    private final Context previous;
-    private final Context current;
-
-    private SpanInScope(Span span) {
-      current = ContextUtils.withValue(span);
-      previous = current.attach();
-    }
-
-    public static SpanInScope create(Span span) {
-      return new SpanInScope(span);
-    }
-
-    @Override
-    public void close() {
-      current.detach(previous);
-    }
   }
 }
